Debug: simplify lytics handler
